+++ dirty_cat/datasets/fetching.py
@@ -24,8 +24,8 @@ import pandas as pd
 from pathlib import Path
 from collections import namedtuple
 from typing import Union, Dict, Any
-from distutils.version import LooseVersion
 
+from dirty_cat.utils import Version
 from dirty_cat.datasets.utils import get_data_dir
 
 
@@ -107,7 +107,9 @@ def fetch_openml_dataset(dataset_id: int,
             "Downloading it from OpenML; this might take a while... "
             "If it is interrupted, some files might be invalid/incomplete: "
             "if on the following run, the fetching raises errors, you can try "
-            f"fixing this issue by deleting the directory {data_directory!r}."
+            f"fixing this issue by deleting the directory {data_directory!s}.",
+            UserWarning,
+            stacklevel=2,
         )
         _download_and_write_openml_dataset(dataset_id=dataset_id,
                                            data_directory=data_directory)
@@ -165,7 +167,7 @@ def _download_and_write_openml_dataset(dataset_id: int,
     from sklearn.datasets import fetch_openml
 
     fetch_kwargs = {}
-    if LooseVersion(sklearn.__version__) >= LooseVersion('0.22'):
+    if Version(sklearn.__version__) >= Version('0.22'):
         fetch_kwargs.update({'as_frame': True})
 
     # The ``fetch_openml()`` function returns a Scikit-Learn ``Bunch`` object,
@@ -349,7 +351,8 @@ def fetch_employee_salaries(load_dataframe: bool = True,
                             drop_linked: bool = True,
                             drop_irrelevant: bool = True,
                             ) -> Union[DatasetAll, DatasetInfoOnly]:
-    """Fetches the employee_salaries dataset.
+    """Fetches the employee_salaries dataset, available at
+    https://openml.org/d/42125
 
     Description of the dataset:
     > Annual salary information including gross pay and overtime pay for all
@@ -400,7 +403,8 @@ def fetch_employee_salaries(load_dataframe: bool = True,
 
 def fetch_road_safety(load_dataframe: bool = True,
                       ) -> Union[DatasetAll, DatasetInfoOnly]:
-    """Fetches the road safety dataset.
+    """Fetches the road safety dataset, available at
+    https://openml.org/d/42803
 
     Description of the dataset:
     > Data reported to the police about the circumstances of personal injury
@@ -432,7 +436,8 @@ def fetch_road_safety(load_dataframe: bool = True,
 
 def fetch_medical_charge(load_dataframe: bool = True
                          ) -> Union[DatasetAll, DatasetInfoOnly]:
-    """Fetches the medical charge dataset.
+    """Fetches the medical charge dataset, available at
+    https://openml.org/d/42720
 
     Description of the dataset:
     > The Inpatient Utilization and Payment Public Use File (Inpatient PUF)
@@ -469,7 +474,8 @@ def fetch_medical_charge(load_dataframe: bool = True
 
 def fetch_midwest_survey(load_dataframe: bool = True
                          ) -> Union[DatasetAll, DatasetInfoOnly]:
-    """Fetches the midwest survey dataset.
+    """Fetches the midwest survey dataset, available at
+    https://openml.org/d/42805
 
     Description of the dataset:
     > Survey to know if people self-identify as Midwesterners.
@@ -499,7 +505,8 @@ def fetch_midwest_survey(load_dataframe: bool = True
 
 def fetch_open_payments(load_dataframe: bool = True
                         ) -> Union[DatasetAll, DatasetInfoOnly]:
-    """Fetches the open payments dataset.
+    """Fetches the open payments dataset, available at
+    https://openml.org/d/42738
 
     Description of the dataset:
     > Payments given by healthcare manufacturing companies to medical doctors
@@ -531,7 +538,8 @@ def fetch_open_payments(load_dataframe: bool = True
 
 def fetch_traffic_violations(load_dataframe: bool = True
                              ) -> Union[DatasetAll, DatasetInfoOnly]:
-    """Fetches the traffic violations dataset.
+    """Fetches the traffic violations dataset, available at
+    https://openml.org/d/42132
 
     Description of the dataset:
     > This dataset contains traffic violation information from all electronic
@@ -565,7 +573,8 @@ def fetch_traffic_violations(load_dataframe: bool = True
 
 def fetch_drug_directory(load_dataframe: bool = True
                          ) -> Union[DatasetAll, DatasetInfoOnly]:
-    """Fetches the drug directory dataset.
+    """Fetches the drug directory dataset, available at
+    https://openml.org/d/43044
 
     Description of the dataset:
     > Product listing data submitted to the U.S. FDA for all unfinished,